Add function validateUser and registerUser
