+++ js/utils/api.mjs
@@ -1,3 +1,3 @@
 export const baseApi = 'https://api.noroff.dev/api/v1';
 export const loginUrl = `${baseApi}/social/auth/login`;
-// export const registerUrl = `${baseApi}/social/auth/register`;
+export const registerUrl = `${baseApi}/social/auth/register`;
