+++ pages/feed/index.html
@@ -12,7 +12,7 @@
 <body class="body-profile bg-primary">
   <nav class="navbar navbar-expand-md navbar-light bg-light">
     <div class="container-sm ">
-      <a class="navbar-brand" href="/feed">MyClass</a>
+      <a class="navbar-brand" href="/">MyClass</a>
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
         aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
@@ -20,10 +20,10 @@
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav me-auto mb-2 mb-md-0">
           <li class="nav-item">
-            <a class="nav-link" aria-current="page" href="/profile">Profile</a>
+            <a class="nav-link" aria-current="page" href="../profile">Profile</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link active" href="/feed">Feed</a>
+            <a class="nav-link active" href="../feed">Feed</a>
           </li>
           <li class="nav-item">
             <a class="nav-link log-out" href="/"><i class="fas fa-sign-out-alt"></i></a></i>
