+++ pages/profile/index.html
@@ -12,7 +12,7 @@
 <body class="body-profile">
   <nav class="navbar navbar-expand-md navbar-light bg-light">
     <div class="container-sm ">
-      <a class="navbar-brand" href="/feed">MyClass</a>
+      <a class="navbar-brand" href="/">MyClass</a>
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
         aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
@@ -20,10 +20,10 @@
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav me-auto mb-2 mb-md-0">
           <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="/profile">Profile</a>
+            <a class="nav-link active" aria-current="page" href="../profile">Profile</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="/feed">Feed</a>
+            <a class="nav-link" href="../feed">Feed</a>
           </li>
           <li class="nav-item">
             <a class="nav-link log-out" href="/"><i class="fas fa-sign-out-alt"></i></a></i>
@@ -50,7 +50,7 @@
     <div class="container-md bg-secondary py-2">
       <div class="row flex-column align-items-center flex-sm-row justify-content-sm-center">
         <div class="col-6 col-sm-3 col-md-2 img-wrapper d-flex justify-content-center d-sm-flex justify-content-sm-end">
-          <img class="border border-4 rounded-circle" src="../images/pexels-antonio-lorenzana-bermejo-14577571.jpg"
+          <img class="border border-4 rounded-circle" src="../../images/pexels-antonio-lorenzana-bermejo-14577571.jpg"
             alt="antonio">
         </div>
         <div class="col-6 col-sm-3 col-md-2 text-light text-center  d-sm-flex flex-sm-column align-items-sm-start user">
