+++ pages/register/index.html
@@ -5,8 +5,8 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>My Class || Authentication</title>
-  <link rel="stylesheet" href="dist/css/index.css">
+  <title>My Class || Register</title>
+  <link rel="stylesheet" href="../../dist/css/index.css">
 </head>
 
 <body class="body-authentication">
@@ -18,10 +18,17 @@
           <div class="col-12 text-white d-flex flex-column justify-content-center align-items-center">
             <h1 class="h1">MyClass</h1>
             <p class="mb-0">Welcome to MyClass Media.</p>
-            <p>Login and enjoy!</p>
+            <p>Register user profile!</p>
           </div>
           <div class="col-12 d-flex flex-column justify-content-center align-items-center p-4">
-            <form action="/pages/profile/" class="col-12">
+            <form action="" class="col-12">
+              <div class="form-group text-white">
+                <label for="userName" class="form-label">User name</label>
+                <input type="text" class="form-control bg-light" id="userName" placeholder="Type name">
+                <p class="name-error d-none">Field can't be empty, must not contain punctuation symbols apart from
+                  underscore
+                  (_).</p>
+              </div>
               <div class="form-group text-white mt-2">
                 <label class="py-1" for="exampleInputEmail1">Email address</label>
                 <input type="email" class="form-control bg-light" id="email" aria-describedby="emailHelp"
@@ -34,10 +41,8 @@
                   required>
                 <p class="password-error d-none">Password must be min 8 characters long</p>
               </div>
-              <button type="submit" class="btn btn-secondary mt-3 text-white col-12">Log
-                in</button>
-              <p class="text-warning py-3">Don't have an account? <a href="pages/register/" class="text-warning">Sign
-                  up!</a></p>
+              <button id="btn-sign-up" type="submit" class="btn btn-secondary mt-3 text-white col-12">Sign up!</button>
+              <p class="text-warning py-3">Are you registered? <a href="../../" class="text-warning">Log in!</a></p>
             </form>
           </div>
         </div>
@@ -47,7 +52,7 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz"
     crossorigin="anonymous"></script>
-  <script type="module" src="js/authentication/index.mjs"></script>
+  <script type="module" src="../../js/authentication/index.mjs"></script>
 </body>
 
 </html>
